add intr mode for host dsp sci

+++ INCDIR/Hboard/boarddrv.h
@@ -66,8 +66,8 @@ UINT8  UartCharGetTimeout(UINT32 timeout, int *result);
 void UartPutbuffer(UINT8 *s);
 
 UINT8 UartCharGetTimeout_B(UINT32 timeout, int *result);
-int UartWrite_B(UINT8 *buffer, UINT16 length);
-int UartRead_B(UINT8* buffer, UINT16 length);
+INT32 UartWrite_B(UINT8 *buffer, UINT32 length);
+INT32 UartRead_B(UINT8* buffer, UINT32 length);
 
 //spi
 void SpiWriteData(UINT16 data);
